+++ src/ragas/metrics/__init__.py
@@ -1,9 +1,4 @@
 from ragas.metrics._answer_correctness import AnswerCorrectness, answer_correctness
-from ragas.metrics._answer_relevance import (
-    AnswerRelevancy,
-    ResponseRelevancy,
-    answer_relevancy,
-)
 from ragas.metrics._answer_similarity import (
     AnswerSimilarity,
     SemanticSimilarity,
@@ -53,6 +48,11 @@ from ragas.metrics._multi_modal_relevance import (
     multimodal_relevance,
 )
 from ragas.metrics._noise_sensitivity import NoiseSensitivity
+from ragas.metrics._response_relevance import (
+    AnswerRelevancy,
+    ResponseRelevancy,
+    answer_relevancy,
+)
 from ragas.metrics._rouge_score import RougeScore
 from ragas.metrics._sql_semantic_equivalence import LLMSQLEquivalence
 from ragas.metrics._string import (
